requirements.txt
fabric
pyyaml
cryptography
oauth2client
google-cloud-secret-manager
streamlit
streamlit-navigation-bar
pandas
numpy
psutil  # used for system metrics logging in mlflow
mlflow
mlflow[extras]
mlserver
# pyenv  VIA BREW/BASH 
pycryptodome
# jose # JWT

sqlmodel

grpcio
pymilvus

langchain
langchain-openai

feast[redis]

dacite  # loading and saving of nested dataclasses